backend/requirements.txt
aiofiles==23.2.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
cssselect2==0.7.0
ecdsa==0.18.0
fastapi==0.100.1
fonttools==4.42.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
logger==1.4
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.0
opencv-python==4.8.0.76
pandas==2.1.2
passlib==1.7.4
pdfkit==1.0.0
Pillow==10.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydyf==0.7.0
PyPDF2==3.0.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.27.0
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.1
weasyprint==59.0
webencodings==0.5.1
XlsxWriter==3.1.9
zopfli==0.2.2
